pyproject.toml
[project]
name = "entropy-data"
version = "0.2.8"
description = "CLI for Entropy Data"
requires-python = ">=3.12"
license = "MIT"
authors = [{ name = "Entropy Data", email = "support@entropy-data.com" }]
dependencies = [
    "typer>=0.15.0,<1.0",
    "rich>=14.0,<16.0",
    "pydantic>=2.10,<3.0",
    "requests>=2.32,<3.0",
    "tomli_w>=1.0,<2.0",
    "pyyaml>=6.0,<7.0",
    "python-dotenv>=1.0,<2.0",
]

[project.urls]
Homepage = "https://entropy-data.com"
Documentation = "https://docs.entropy-data.com"
Repository = "https://github.com/entropy-data/entropy-data-cli"

[project.scripts]
entropy-data = "entropy_data.cli:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "responses>=0.25",
    "ruff>=0.8",
    "pre-commit>=3.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/entropy_data"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
extend-select = ["I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
